renaming modules + fixed lint
