+++ lambdas/batch_processing/sample_data/patient_data3.csv
@@ -1,2 +1,2 @@
-NHS_NUMBER,PERSON_FORENAME,PERSON_SURNAME,PERSON_DOB,PERSON_GENDER_CODE,PERSON_POSTCODE,DATE_AND_TIME,SITE_CODE,ACTION_FLAG,PERFORMING_PROFESSIONAL_FORENAME,PERFORMING_PROFESSIONAL_SURNAME,PERFORMING_PROFESSIONAL_BODY_REG_URI,PERFORMING_PROFESSIONAL_BODY_REG_CODE,RECORDED_DATE,PRIMARY_SOURCE,REPORT_ORIGIN,NOT_GIVEN,VACCINATION_PROCEDURE_CODE,VACCINATION_PROCEDURE_TERM,VACCINATION_SITUATION_CODE,REASON_NOT_GIVEN_CODE,REASON_NOT_GIVEN_TERM,DOSE_SEQUENCE,VACCINE_PRODUCT_CODE,VACCINE_PRODUCT_TERM,VACCINE_MANUFACTURER,BATCH_NUMBER,EXPIRY_DATE,ROUTE_OF_VACCINATION_CODE,DOSE_AMOUNT,DOSE_UNIT_CODE,INDICATION_CODE,CONSENT_FOR_TREATMENT_CODE,SUBMITTED_TIMESTAMP
-123 456 7899,first,test,1952-05-05,2,M242 2HYY,2020-12-14T10:08:15+00,B0C4P,completed,testforename,testsurname,https://fhir.hl7.org.uk/Id/nmc-number,reg_code,2022-05-05,true,X99999,empty,VACCINATION_PROCEDURE_CODE,VACCINATION_PROCEDURE_TERM,VACCINATION_SITUATION_CODE,REASON_NOT_GIVEN_CODE,REASON_NOT_GIVEN_TERM,2,VACCINE_PRODUCT_CODE,VACCINE_PRODUCT_TERM,VACCINE_MANUFACTURER,BATCH_NUMBER,1952-05-05,ROUTE_OF_VACCINATION_CODE,4.1,DOSE_UNIT_CODE,INDICATION_CODE,CONSENT_FOR_TREATMENT_CODE,2020-12-14T10:08:15+00
+NHS_NUMBER,PERSON_FORENAME,PERSON_SURNAME,PERSON_DOB,PERSON_GENDER_CODE,PERSON_POSTCODE,DATE_AND_TIME,SITE_CODE,ACTION_FLAG,PERFORMING_PROFESSIONAL_FORENAME,PERFORMING_PROFESSIONAL_SURNAME,PERFORMING_PROFESSIONAL_BODY_REG_URI,PERFORMING_PROFESSIONAL_BODY_REG_CODE,RECORDED_DATE,PRIMARY_SOURCE,REPORT_ORIGIN,NOT_GIVEN,VACCINATION_PROCEDURE_CODE,VACCINATION_PROCEDURE_TERM,VACCINATION_SITUATION_CODE,REASON_NOT_GIVEN_CODE,REASON_NOT_GIVEN_TERM,DOSE_SEQUENCE,VACCINE_PRODUCT_CODE,VACCINE_PRODUCT_TERM,VACCINE_MANUFACTURER,BATCH_NUMBER,EXPIRY_DATE,ROUTE_OF_VACCINATION_CODE,DOSE_AMOUNT,DOSE_UNIT_CODE,INDICATION_CODE,CONSENT_FOR_TREATMENT_CODE,SUBMITTED_TIMESTAMP,LOCATION_CODE,REDUCE_VALIDATION_CODE,REDUCE_VALIDATION_REASON
+,first,test,1952-05-05,2,M242 2HYY,2020-12-14T10:08:15+00,B0C4P,completed,testforename,testsurname,https://fhir.hl7.org.uk/Id/nmc-number,reg_code,2022-05-05,true,X99999,empty,VACCINATION_PROCEDURE_CODE,VACCINATION_PROCEDURE_TERM,VACCINATION_SITUATION_CODE,REASON_NOT_GIVEN_CODE,REASON_NOT_GIVEN_TERM,2,VACCINE_PRODUCT_CODE,VACCINE_PRODUCT_TERM,VACCINE_MANUFACTURER,BATCH_NUMBER,1952-05-05,ROUTE_OF_VACCINATION_CODE,4.1,DOSE_UNIT_CODE,INDICATION_CODE,CONSENT_FOR_TREATMENT_CODE,2020-12-14T10:08:15+00,ODS123,True,REDUCE_VALIDATION_REASON
